update: aurora.py
added command at the beginning of the code that makes it execute the program without the need to add a manifest.
added command to execute a set execution policy at the beginning of the code to allow the execution of powershell scripts.
Removed the commands file from the repository.
Aurora will no longer have .exe binaries due to false virus positives, but we are already working on a solution.

It is now not necessary to click on the first item in the list to enable the execution of scripts, however, if it does not work for you, you may need to click on the first item in the list to enable the execution of powershell scripts.